requirements/lock_raw.txt
absl-py==2.4.0
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
audioread==3.1.0
av==12.3.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
ctranslate2==4.7.1
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.1
fastapi==0.136.1
faster-whisper==1.0.3
filelock==3.29.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.4.0
greenlet==3.5.0
h11==0.16.0
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.10.2
llvmlite==0.47.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mediapipe==0.10.9
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.65.1
numpy==1.26.4
onnxruntime==1.23.2
opencv-contrib-python==4.13.0.92
packaging==26.2
pandas==2.2.2
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
protobuf==3.20.3
psutil==7.2.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyparsing==3.3.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.27
pytz==2026.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.13.1
six==1.17.0
sounddevice==0.5.5
soundfile==0.12.1
soxr==1.1.0
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
tqdm==4.66.4
transformers==4.41.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
